+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.6.0 (unreleased)
 ------------------
 
+- #2563 Move Dynamic Analysis Specs to Setup folder
 - #2554 Migrate AttachmentTypes to Dexterity
 - #2572 Move Sample Containers to Senaite setup folder
 - #2571 Move InstrumentLocations to Senaite setup folder